Refatora o projeto para maior organização e simplicidade

+++ BACKEND/BugOut-Backend/src/main/java/br/unicap/bugout/user/UserController.java
@@ -1,17 +1,22 @@
 package br.unicap.bugout.user;
 
+import br.unicap.bugout.shared.MessageDTO;
+import br.unicap.bugout.shared.NoModificationException;
+import br.unicap.bugout.shared.AdminUserCannotBeModifiedException;
+import br.unicap.bugout.user.exceptions.UserAlreadyExistsException;
+import br.unicap.bugout.user.exceptions.UserNotFoundException;
+import br.unicap.bugout.user.model.User;
+import br.unicap.bugout.user.model.UserDTO;
+import br.unicap.bugout.user.model.UserMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-
-import java.util.List;
-
-import org.mapstruct.MappingTarget;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
-import br.unicap.bugout.shared.MessageDTO;
+import javax.validation.Valid;
+import java.util.List;
 
 @Slf4j
 @RestController
@@ -26,48 +31,68 @@ public class UserController {
     private final UserMapper mapper;
 
 
+    @GetMapping("/{id}")
+    public ResponseEntity<UserDTO> getById(@PathVariable Long id) {
+        log.info("{}/Get By ID - ID {}", PATH, id);
+
+        User user = service.getById(id);
+        return ResponseEntity.ok(mapper.toDTO(user));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<UserDTO>> getAll() {
+        log.info("{}/Get All", PATH);
+
+        List<User> list = service.getAll();
+        return ResponseEntity.ok(mapper.toDTOs(list));
+    }
+
     @PostMapping
     @Transactional
-    public ResponseEntity<UserDTO> create(@RequestBody UserDTO dto) {
+    public ResponseEntity<UserDTO> create(@Valid @RequestBody UserDTO dto) {
         log.info("{}/Create", PATH);
 
-        boolean exists = service.exists(dto.getUsername(), dto.getEmail());
-        if (exists)
-            throw new UserAlreadyExistsException();
+        boolean exists = service.exists(dto.getUsername().trim(), dto.getEmail().trim());
+        if (exists) throw new UserAlreadyExistsException();
 
         User created = service.save(mapper.toEntity(dto));
         return new ResponseEntity<>(mapper.toDTO(created), HttpStatus.CREATED);
     }
 
-    @GetMapping
-    public ResponseEntity<UserDTO> getUser(@RequestParam Long id){
-        User user = service.getUser(id);
-
-        return new ResponseEntity<>(mapper.toDTO(user), HttpStatus.OK);
-    }
-
-    @GetMapping("/all")
-    public ResponseEntity<List<UserDTO>> getAllUsers(){
-        List<User> allUsers = service.getAllUsers();
-
-        return new ResponseEntity<>(mapper.toDTOs(allUsers), HttpStatus.OK);
-    }
-
-    @PutMapping
     @Transactional
-    public ResponseEntity<UserDTO> updateUser(@RequestBody UserDTO dto){
-        boolean exists = service.exists(dto.getUsername(), dto.getEmail());
-        User user = new User();
-        if (!exists)
-            throw new UserAlreadyExistsException();
+    @PutMapping("/{id}")
+    public ResponseEntity<UserDTO> update(@PathVariable Long id, @Valid @RequestBody UserDTO dto) {
+        log.info("{}/Update - ID {}", PATH, id);
 
-        User userUpdated = service.updateUser(mapper.updateEntity(dto, user));
-        return new ResponseEntity<>(mapper.toDTO(userUpdated), HttpStatus.OK);
+        if (service.isAdmin(id)) throw new AdminUserCannotBeModifiedException();
+
+        User user = service.getById(id);
+
+
+        String username = dto.getUsername().trim();
+        String email = dto.getEmail().trim();
+
+        if (user.getUsername().equalsIgnoreCase(username) && user.getEmail().equalsIgnoreCase(email))
+            throw new NoModificationException();
+
+        boolean exists = service.existsOtherThanSelf(id, username, email);
+        if (exists) throw new UserAlreadyExistsException();
+
+
+        User updated = service.save(mapper.updateEntity(dto, user));
+        return ResponseEntity.ok(mapper.toDTO(updated));
     }
 
+    @Transactional
     @DeleteMapping("/{id}")
-    @Transactional
-    public ResponseEntity<MessageDTO> deleteUser(@PathVariable Long id){
+    public ResponseEntity<MessageDTO> delete(@PathVariable Long id) {
+        log.info("{}/Delete - ID {}", PATH, id);
+
+        if (service.isAdmin(id)) throw new AdminUserCannotBeModifiedException();
+
+        boolean exists = service.existsById(id);
+        if (!exists) throw new UserNotFoundException();
+
         service.deleteById(id);
         return ResponseEntity.ok(new MessageDTO("Usuário deletado com sucesso!"));
     }
